Offer up a Config; provide Sub; provide Default...
With the Config we could expand to more options in the future. The
NilWriter had to be made so we can differentiate between
nil-I-want-the-default and NilWriter-I-want-no-output.

The Sub let's you make a Flog based on another flog (with an extra
sub-name for example).

The Default Flog now has mirrored top-level convenience functions.